src/main/java/com/github/havlli/EventPilot/session/SessionRedisAccessService.java
package com.github.havlli.EventPilot.session;

import org.springframework.data.redis.core.StringRedisTemplate;
import org.springframework.stereotype.Repository;

@Repository
public class SessionRedisAccessService implements SessionDAO {

    private final StringRedisTemplate redisTemplate;

    public SessionRedisAccessService(StringRedisTemplate redisTemplate) {
        this.redisTemplate = redisTemplate;
    }
    @Override
    public void save(String key, String value) {
        redisTemplate.opsForValue().set(key, value);
    }

    @Override
    public Boolean exists(String key) {
        return redisTemplate.hasKey(key);
    }

    @Override
    public void remove(String key) {
        redisTemplate.delete(key);
    }

    @Override
    public void clear() {
        redisTemplate.delete(redisTemplate.keys("*"));
    }
}
